requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
Deprecated==1.2.13
Django==3.2.13
django-bootstrap4==22.1
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
importlib-metadata==2.1.3
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.2.0
wrapt==1.14.1
zipp==3.8.0
